Add outbound CallChannel support to Kotlin/Native
